Creation of DAO Interfaces
In this commit, the DepartmentDao and SellerDao interfaces have been added to the project, following the DAO (Data Access Object) design pattern. These interfaces define the methods that will be implemented to perform operations related to departments and sellers in the database, providing an effective abstraction for data access and promoting separation of concerns in the code.